requirements.txt
asgiref==3.7.2
attrs==23.2.0
black==24.2.0
click==8.1.7
Django==4.2.7
django-filter==23.5
django-mail-templated==2.6.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
flake8==7.0.0
importlib-metadata==7.0.1
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
Markdown==3.5.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.2.0
pycodestyle==2.11.1
pyflakes==3.2.0
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
rpds-py==0.17.1
setuptools==68.2.2
sqlparse==0.4.4
swagger-spec-validator==3.0.3
typing_extensions==4.8.0
uritemplate==4.1.1
wheel==0.41.2
zipp==3.17.0
